Fixed a few small bugs
- Authentication on `Cert` fail when verbose error
- Rebased `access_point_shell.py`
